move Bundle Locale tests to BundleTest
