+++ gto-support-util/src/test/java/org/ccci/gto/android/common/util/os/BundleUtilsTest.kt
@@ -2,7 +2,6 @@ package org.ccci.gto.android.common.util.os
 
 import android.os.Bundle
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import java.util.Locale
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
 import org.junit.Test
@@ -55,19 +54,5 @@ class BundleUtilsTest {
         assertEquals(TestEnum.VALUE1, bundle.getEnum(KEY1, TestEnum.DEFVALUE))
         assertEquals(TestEnum.DEFVALUE, bundle.getEnum(KEY2, TestEnum.DEFVALUE))
         assertEquals(TestEnum.DEFVALUE, bundle.getEnum(KEY3, TestEnum.DEFVALUE))
-    }
-
-    @Test
-    fun verifyPutLocaleAndGetLocale() {
-        val bundle = Bundle().apply {
-            putLocale(KEY1, Locale.ENGLISH)
-            putLocale(KEY2, Locale.FRENCH)
-            putLocale(KEY3, null)
-        }
-
-        assertEquals(3, bundle.size())
-        assertEquals(Locale.ENGLISH, bundle.getLocale(KEY1))
-        assertEquals(Locale.FRENCH, bundle.getLocale(KEY2))
-        assertNull(bundle.getLocale(KEY3))
     }
 }
