+++ gto-support-util/src/test/kotlin/org/ccci/gto/android/common/util/os/BundleTest.kt
@@ -3,6 +3,7 @@ package org.ccci.gto.android.common.util.os
 import android.graphics.Point
 import android.os.Bundle
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import java.util.Locale
 import org.junit.Assert.assertArrayEquals
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
@@ -12,10 +13,27 @@ import org.robolectric.annotation.Config
 
 private const val KEY1 = "key1"
 private const val KEY2 = "key2"
+private const val KEY3 = "key3"
 
 @RunWith(AndroidJUnit4::class)
 class BundleTest {
-    // region ParcelableArray
+    // region Locales
+    @Test
+    fun verifyPutLocaleAndGetLocale() {
+        val bundle = Bundle().apply {
+            putLocale(KEY1, Locale.ENGLISH)
+            putLocale(KEY2, Locale.FRENCH)
+            putLocale(KEY3, null)
+        }
+
+        assertEquals(3, bundle.size())
+        assertEquals(Locale.ENGLISH, bundle.getLocale(KEY1))
+        assertEquals(Locale.FRENCH, bundle.getLocale(KEY2))
+        assertNull(bundle.getLocale(KEY3))
+    }
+    // endregion Locales
+
+    // region ParcelableArrays
     @Test
     fun verifyGetParcelableArray() {
         val points = arrayOf(Point(0, 0), null, Point(1, 1))
@@ -43,5 +61,5 @@ class BundleTest {
         assertArrayEquals(points, resp)
         assertNull(bundle.getTypedParcelableArray<Point>(KEY2))
     }
-    // endregion ParcelableArray
+    // endregion ParcelableArrays
 }
